Node add_child method:
Takes reference of Node self, the address of Reference Count to Self, and the child to add to self.

child parameter makes sense. Reference Count of child now gives up ownership of itself to the parent.

We need the self_ref in the parameters because without the Rc to self, the child cannot create a shared reference of the parent.
- It needs the address of the Rc to Self to create its own weak reference to Self based on that originating address of Rc to Self.
- The child's reference to the parent: self cannot simply be a regular reference &self because we'd have to use lifetimes for &self. lifetimes guarantee that the reference to parent will live as long as the child, and while this is the case for our example, it isn't always the case. Also, smart pointers (references) are much more intuitive and safe for shared references.